requirements.txt
boto==2.38.0
django==1.8.7
django-storages==1.1.8
django-spirit==0.4.5
python-social-auth==0.2.7
psycopg2==2.6.1
